Collapse .h and .cc files in DQMOffline/CalibCalo and reorganize the files
