+++ DQMOffline/CalibCalo/plugins/DQMHcalIterativePhiSymAlCaReco.cc
@@ -10,24 +10,69 @@
 
 #include "FWCore/Framework/interface/Event.h"
 #include "FWCore/Framework/interface/EventSetup.h"
+#include "FWCore/Framework/interface/Frameworkfwd.h"
+#include "FWCore/Framework/interface/MakerMacros.h"
+#include "FWCore/MessageLogger/interface/MessageLogger.h"
 #include "FWCore/ParameterSet/interface/ParameterSet.h"
 #include "FWCore/ServiceRegistry/interface/Service.h"
-
-#include "FWCore/MessageLogger/interface/MessageLogger.h"
+#include "FWCore/ParameterSet/interface/ConfigurationDescriptions.h"
+#include "FWCore/ParameterSet/interface/ParameterSet.h"
 
 // DQM include files
 
 #include "DQMServices/Core/interface/DQMStore.h"
+#include "DQMServices/Core/interface/DQMOneEDAnalyzer.h"
 
 // work on collections
 
 #include "DataFormats/HcalDetId/interface/HcalDetId.h"
 #include "DataFormats/HcalDetId/interface/HcalSubdetector.h"
 #include "DataFormats/HcalRecHit/interface/HcalRecHitCollections.h"
+#include "DataFormats/FEDRawData/interface/FEDRawDataCollection.h"
+
 #include "Geometry/CaloGeometry/interface/CaloCellGeometry.h"
 #include "Geometry/CaloGeometry/interface/CaloSubdetectorGeometry.h"
 
-#include "DQMOffline/CalibCalo/interface/DQMHcalIterativePhiSymAlCaReco.h"
+#include <string>
+
+class DQMHcalIterativePhiSymAlCaReco : public DQMOneEDAnalyzer<> {
+public:
+  DQMHcalIterativePhiSymAlCaReco(const edm::ParameterSet &);
+  ~DQMHcalIterativePhiSymAlCaReco() override;
+
+  static void fillDescriptions(edm::ConfigurationDescriptions &descriptions);
+
+protected:
+  //  void beginRun(const edm::Run& r, const edm::EventSetup& c);
+  void bookHistograms(DQMStore::IBooker &, edm::Run const &, edm::EventSetup const &) override;
+  void analyze(const edm::Event &e, const edm::EventSetup &c) override;
+
+  void dqmEndRun(const edm::Run &r, const edm::EventSetup &c) override {}
+
+private:
+  static constexpr int maxDepth_ = 7;
+  //
+  // Monitor elements
+  //
+  MonitorElement *hiDistr2D_[maxDepth_];
+
+  MonitorElement *hiDistrHBHEsize1D_;
+  MonitorElement *hiDistrHFsize1D_;
+  MonitorElement *hiDistrHOsize1D_;
+
+  std::string folderName_;
+  int hiDistr_y_nbin_;
+  int hiDistr_x_nbin_;
+  double hiDistr_y_min_;
+  double hiDistr_y_max_;
+  double hiDistr_x_min_;
+  double hiDistr_x_max_;
+  /// object to monitor
+
+  edm::EDGetTokenT<HBHERecHitCollection> tok_hbhe_;
+  edm::EDGetTokenT<HORecHitCollection> tok_ho_;
+  edm::EDGetTokenT<HFRecHitCollection> tok_hf_;
+};
 
 // ******************************************
 // constructors
@@ -148,3 +193,5 @@ void DQMHcalIterativePhiSymAlCaReco::analyze(const edm::Event &iEvent, const edm
   }
 
 }  // analyze
+
+DEFINE_FWK_MODULE(DQMHcalIterativePhiSymAlCaReco);
