+++ DQMOffline/CalibCalo/plugins/DQMSourcePi0.cc
@@ -1,49 +1,328 @@
 #include "FWCore/Framework/interface/Event.h"
 #include "FWCore/Framework/interface/EventSetup.h"
+#include "FWCore/Framework/interface/Frameworkfwd.h"
+#include "FWCore/Framework/interface/MakerMacros.h"
+#include "FWCore/MessageLogger/interface/MessageLogger.h"
 #include "FWCore/ParameterSet/interface/ParameterSet.h"
 #include "FWCore/ServiceRegistry/interface/Service.h"
+#include "FWCore/Utilities/interface/InputTag.h"
 
 // DQM include files
 
+#include "DQMServices/Core/interface/DQMEDAnalyzer.h"
 #include "DQMServices/Core/interface/DQMStore.h"
 
 // work on collections
+#include "DataFormats/DetId/interface/DetId.h"
 #include "DataFormats/EcalDetId/interface/EBDetId.h"
 #include "DataFormats/EcalDetId/interface/EEDetId.h"
 #include "DataFormats/EcalRecHit/interface/EcalRecHitCollections.h"
-
 #include "DataFormats/EcalRecHit/interface/EcalRecHit.h"
-
-#include "DQMOffline/CalibCalo/src/DQMSourcePi0.h"
-
-#include "DataFormats/DetId/interface/DetId.h"
-#include "DataFormats/EcalDetId/interface/EBDetId.h"
-#include "DataFormats/EcalRecHit/interface/EcalRecHitCollections.h"
 #include "DataFormats/EgammaReco/interface/BasicCluster.h"
 #include "DataFormats/Math/interface/Point3D.h"
-
 #include "DataFormats/Common/interface/Handle.h"
-#include "FWCore/Framework/interface/ESHandle.h"
-#include "FWCore/Framework/interface/Event.h"
-#include "FWCore/Framework/interface/EventSetup.h"
-#include "FWCore/MessageLogger/interface/MessageLogger.h"
-
 #include "DataFormats/RecoCandidate/interface/RecoEcalCandidate.h"
 #include "DataFormats/RecoCandidate/interface/RecoEcalCandidateFwd.h"
 
+// Geometry
+#include "Geometry/Records/interface/CaloTopologyRecord.h"
+#include "Geometry/CaloGeometry/interface/CaloCellGeometry.h"
+#include "Geometry/CaloGeometry/interface/CaloGeometry.h"
+#include "Geometry/CaloGeometry/interface/CaloSubdetectorGeometry.h"
+#include "Geometry/CaloTopology/interface/CaloSubdetectorTopology.h"
+#include "Geometry/CaloTopology/interface/CaloTopology.h"
+#include "Geometry/CaloTopology/interface/EcalBarrelTopology.h"
+#include "Geometry/CaloTopology/interface/EcalEndcapTopology.h"
+#include "Geometry/Records/interface/IdealGeometryRecord.h"
+#include "RecoEcal/EgammaCoreTools/interface/PositionCalc.h"
+
 #include "TVector3.h"
 
-#define TWOPI 6.283185308
+// Less than operator for sorting EcalRecHits according to energy.
+inline bool ecalRecHitGreater(EcalRecHit x, EcalRecHit y) { return (x.energy() > y.energy()); }
 
-using namespace std;
-using namespace edm;
+class DQMSourcePi0 : public DQMEDAnalyzer {
+public:
+  DQMSourcePi0(const edm::ParameterSet &);
+  ~DQMSourcePi0() override;
+
+protected:
+  void bookHistograms(DQMStore::IBooker &, edm::Run const &, edm::EventSetup const &) override;
+  void analyze(const edm::Event &e, const edm::EventSetup &c) override;
+
+  void convxtalid(int &, int &);
+  int diff_neta_s(int, int);
+  int diff_nphi_s(int, int);
+
+private:
+  int eventCounter_;
+  PositionCalc posCalculator_;
+
+  /// Distribution of rechits in iPhi (pi0)
+  MonitorElement *hiPhiDistrEBpi0_;
+
+  /// Distribution of rechits in ix EE  (pi0)
+  MonitorElement *hiXDistrEEpi0_;
+
+  /// Distribution of rechits in iPhi (eta)
+  MonitorElement *hiPhiDistrEBeta_;
+
+  /// Distribution of rechits in ix EE  (eta)
+  MonitorElement *hiXDistrEEeta_;
+
+  /// Distribution of rechits in iEta (pi0)
+  MonitorElement *hiEtaDistrEBpi0_;
+
+  /// Distribution of rechits in iy EE (pi0)
+  MonitorElement *hiYDistrEEpi0_;
+
+  /// Distribution of rechits in iEta (eta)
+  MonitorElement *hiEtaDistrEBeta_;
+
+  /// Distribution of rechits in iy EE (eta)
+  MonitorElement *hiYDistrEEeta_;
+
+  /// Energy Distribution of rechits EB (pi0)
+  MonitorElement *hRechitEnergyEBpi0_;
+
+  /// Energy Distribution of rechits EE (pi0)
+  MonitorElement *hRechitEnergyEEpi0_;
+
+  /// Energy Distribution of rechits EB (eta)
+  MonitorElement *hRechitEnergyEBeta_;
+
+  /// Energy Distribution of rechits EE (eta)
+  MonitorElement *hRechitEnergyEEeta_;
+
+  /// Distribution of total event energy EB (pi0)
+  MonitorElement *hEventEnergyEBpi0_;
+
+  /// Distribution of total event energy EE (pi0)
+  MonitorElement *hEventEnergyEEpi0_;
+
+  /// Distribution of total event energy EB (eta)
+  MonitorElement *hEventEnergyEBeta_;
+
+  /// Distribution of total event energy EE (eta)
+  MonitorElement *hEventEnergyEEeta_;
+
+  /// Distribution of number of RecHits EB (pi0)
+  MonitorElement *hNRecHitsEBpi0_;
+
+  /// Distribution of number of RecHits EE (pi0)
+  MonitorElement *hNRecHitsEEpi0_;
+
+  /// Distribution of number of RecHits EB (eta)
+  MonitorElement *hNRecHitsEBeta_;
+
+  /// Distribution of number of RecHits EE (eta)
+  MonitorElement *hNRecHitsEEeta_;
+
+  /// Distribution of Mean energy per rechit EB (pi0)
+  MonitorElement *hMeanRecHitEnergyEBpi0_;
+
+  /// Distribution of Mean energy per rechit EE (pi0)
+  MonitorElement *hMeanRecHitEnergyEEpi0_;
+
+  /// Distribution of Mean energy per rechit EB (eta)
+  MonitorElement *hMeanRecHitEnergyEBeta_;
+
+  /// Distribution of Mean energy per rechit EE (eta)
+  MonitorElement *hMeanRecHitEnergyEEeta_;
+
+  /// Pi0 invariant mass in EB
+  MonitorElement *hMinvPi0EB_;
+
+  /// Pi0 invariant mass in EE
+  MonitorElement *hMinvPi0EE_;
+
+  /// Eta invariant mass in EB
+  MonitorElement *hMinvEtaEB_;
+
+  /// Eta invariant mass in EE
+  MonitorElement *hMinvEtaEE_;
+
+  /// Pt of the 1st most energetic Pi0 photon in EB
+  MonitorElement *hPt1Pi0EB_;
+
+  /// Pt of the 1st most energetic Pi0 photon in EE
+  MonitorElement *hPt1Pi0EE_;
+
+  /// Pt of the 1st most energetic Eta photon in EB
+  MonitorElement *hPt1EtaEB_;
+
+  /// Pt of the 1st most energetic Eta photon in EE
+  MonitorElement *hPt1EtaEE_;
+
+  /// Pt of the 2nd most energetic Pi0 photon in EB
+  MonitorElement *hPt2Pi0EB_;
+
+  /// Pt of the 2nd most energetic Pi0 photon in EE
+  MonitorElement *hPt2Pi0EE_;
+
+  /// Pt of the 2nd most energetic Eta photon in EB
+  MonitorElement *hPt2EtaEB_;
+
+  /// Pt of the 2nd most energetic Eta photon in EE
+  MonitorElement *hPt2EtaEE_;
+
+  /// Pi0 Pt in EB
+  MonitorElement *hPtPi0EB_;
+
+  /// Pi0 Pt in EE
+  MonitorElement *hPtPi0EE_;
+
+  /// Eta Pt in EB
+  MonitorElement *hPtEtaEB_;
+
+  /// Eta Pt in EE
+  MonitorElement *hPtEtaEE_;
+
+  /// Pi0 Iso EB
+  MonitorElement *hIsoPi0EB_;
+
+  /// Pi0 Iso EE
+  MonitorElement *hIsoPi0EE_;
+
+  /// Eta Iso EB
+  MonitorElement *hIsoEtaEB_;
+
+  /// Eta Iso EE
+  MonitorElement *hIsoEtaEE_;
+
+  /// S4S9 of the 1st most energetic pi0 photon
+  MonitorElement *hS4S91Pi0EB_;
+
+  /// S4S9 of the 1st most energetic pi0 photon EE
+  MonitorElement *hS4S91Pi0EE_;
+
+  /// S4S9 of the 1st most energetic eta photon
+  MonitorElement *hS4S91EtaEB_;
+
+  /// S4S9 of the 1st most energetic eta photon EE
+  MonitorElement *hS4S91EtaEE_;
+
+  /// S4S9 of the 2nd most energetic pi0 photon
+  MonitorElement *hS4S92Pi0EB_;
+
+  /// S4S9 of the 2nd most energetic pi0 photon EE
+  MonitorElement *hS4S92Pi0EE_;
+
+  /// S4S9 of the 2nd most energetic eta photon
+  MonitorElement *hS4S92EtaEB_;
+
+  /// S4S9 of the 2nd most energetic eta photon EE
+  MonitorElement *hS4S92EtaEE_;
+
+  /// object to monitor
+  edm::EDGetTokenT<EcalRecHitCollection> productMonitoredEBpi0_;
+  edm::EDGetTokenT<EcalRecHitCollection> productMonitoredEBeta_;
+
+  /// object to monitor
+  edm::EDGetTokenT<EcalRecHitCollection> productMonitoredEEpi0_;
+  edm::EDGetTokenT<EcalRecHitCollection> productMonitoredEEeta_;
+
+  edm::ESGetToken<CaloTopology, CaloTopologyRecord> caloTopoToken_;
+  edm::ESGetToken<CaloGeometry, CaloGeometryRecord> caloGeomToken_;
+
+  int gammaCandEtaSize_;
+  int gammaCandPhiSize_;
+
+  double seleXtalMinEnergy_;
+  double seleXtalMinEnergyEndCap_;
+
+  double clusSeedThr_;
+  int clusEtaSize_;
+  int clusPhiSize_;
+
+  double clusSeedThrEndCap_;
+
+  //// for pi0->gg barrel
+  double selePtGamma_;
+  double selePtPi0_;
+  double seleMinvMaxPi0_;
+  double seleMinvMinPi0_;
+  double seleS4S9Gamma_;
+  double selePi0BeltDR_;
+  double selePi0BeltDeta_;
+  double selePi0Iso_;
+  double ptMinForIsolation_;
+
+  /// for pi0->gg endcap
+  double selePtGammaEndCap_;
+  double selePtPi0EndCap_;
+  double seleMinvMaxPi0EndCap_;
+  double seleMinvMinPi0EndCap_;
+  double seleS4S9GammaEndCap_;
+  double selePi0IsoEndCap_;
+  double selePi0BeltDREndCap_;
+  double selePi0BeltDetaEndCap_;
+  double ptMinForIsolationEndCap_;
+
+  /// for eta->gg barrel
+  double selePtGammaEta_;
+  double selePtEta_;
+  double seleS4S9GammaEta_;
+  double seleS9S25GammaEta_;
+  double seleMinvMaxEta_;
+  double seleMinvMinEta_;
+  double ptMinForIsolationEta_;
+  double seleEtaIso_;
+  double seleEtaBeltDR_;
+  double seleEtaBeltDeta_;
+
+  /// for eta->gg endcap
+  double selePtGammaEtaEndCap_;
+  double seleS4S9GammaEtaEndCap_;
+  double seleS9S25GammaEtaEndCap_;
+  double selePtEtaEndCap_;
+  double seleMinvMaxEtaEndCap_;
+  double seleMinvMinEtaEndCap_;
+  double ptMinForIsolationEtaEndCap_;
+  double seleEtaIsoEndCap_;
+  double seleEtaBeltDREndCap_;
+  double seleEtaBeltDetaEndCap_;
+
+  bool ParameterLogWeighted_;
+  double ParameterX0_;
+  double ParameterT0_barl_;
+  double ParameterT0_endc_;
+  double ParameterT0_endcPresh_;
+  double ParameterW0_;
+
+  std::vector<EBDetId> detIdEBRecHits;
+  std::vector<EcalRecHit> EBRecHits;
+
+  std::vector<EEDetId> detIdEERecHits;
+  std::vector<EcalRecHit> EERecHits;
+
+  /// Monitor every prescaleFactor_ events
+  unsigned int prescaleFactor_;
+
+  /// DQM folder name
+  std::string folderName_;
+
+  /// Write to file
+  bool saveToFile_;
+
+  /// which subdet will be monitored
+  bool isMonEBpi0_;
+  bool isMonEBeta_;
+  bool isMonEEpi0_;
+  bool isMonEEeta_;
+
+  /// Output file name if required
+  std::string fileName_;
+};
+
+#define TWOPI 6.283185308
 
 // ******************************************
 // constructors
 // *****************************************
 
 DQMSourcePi0::DQMSourcePi0(const edm::ParameterSet &ps) : eventCounter_(0) {
-  folderName_ = ps.getUntrackedParameter<string>("FolderName", "HLT/AlCaEcalPi0");
+  folderName_ = ps.getUntrackedParameter<std::string>("FolderName", "HLT/AlCaEcalPi0");
   prescaleFactor_ = ps.getUntrackedParameter<int>("prescaleFactor", 1);
   productMonitoredEBpi0_ =
       consumes<EcalRecHitCollection>(ps.getUntrackedParameter<edm::InputTag>("AlCaStreamEBpi0Tag"));
@@ -62,7 +341,7 @@ DQMSourcePi0::DQMSourcePi0(const edm::ParameterSet &ps) : eventCounter_(0) {
   isMonEEeta_ = ps.getUntrackedParameter<bool>("isMonEEeta", false);
 
   saveToFile_ = ps.getUntrackedParameter<bool>("SaveToFile", false);
-  fileName_ = ps.getUntrackedParameter<string>("FileName", "MonitorAlCaEcalPi0.root");
+  fileName_ = ps.getUntrackedParameter<std::string>("FileName", "MonitorAlCaEcalPi0.root");
 
   clusSeedThr_ = ps.getParameter<double>("clusSeedThr");
   clusSeedThrEndCap_ = ps.getParameter<double>("clusSeedThrEndCap");
@@ -304,7 +583,7 @@ void DQMSourcePi0::bookHistograms(DQMStore::IBooker &ibooker, edm::Run const &ir
 }
 
 //-------------------------------------------------------------
-void DQMSourcePi0::analyze(const Event &iEvent, const EventSetup &iSetup) {
+void DQMSourcePi0::analyze(const edm::Event &iEvent, const edm::EventSetup &iSetup) {
   if (eventCounter_ % prescaleFactor_)
     return;
   eventCounter_++;
@@ -314,7 +593,7 @@ void DQMSourcePi0::analyze(const Event &iEvent, const EventSetup &iSetup) {
   std::vector<EcalRecHit> seeds;
   seeds.clear();
 
-  vector<EBDetId> usedXtals;
+  std::vector<EBDetId> usedXtals;
   usedXtals.clear();
 
   detIdEBRecHits.clear();  //// EBDetId
@@ -387,22 +666,22 @@ void DQMSourcePi0::analyze(const Event &iEvent, const EventSetup &iSetup) {
       // cout<< " Pi0 seeds: "<<seeds.size()<<endl;
 
       int nClus;
-      vector<float> eClus;
-      vector<float> etClus;
-      vector<float> etaClus;
-      vector<float> thetaClus;
-      vector<float> phiClus;
-      vector<EBDetId> max_hit;
+      std::vector<float> eClus;
+      std::vector<float> etClus;
+      std::vector<float> etaClus;
+      std::vector<float> thetaClus;
+      std::vector<float> phiClus;
+      std::vector<EBDetId> max_hit;
 
-      vector<vector<EcalRecHit>> RecHitsCluster;
-      vector<vector<EcalRecHit>> RecHitsCluster5x5;
-      vector<float> s4s9Clus;
-      vector<float> s9s25Clus;
+      std::vector<std::vector<EcalRecHit>> RecHitsCluster;
+      std::vector<std::vector<EcalRecHit>> RecHitsCluster5x5;
+      std::vector<float> s4s9Clus;
+      std::vector<float> s9s25Clus;
 
       nClus = 0;
 
       // Make own simple clusters (3x3, 5x5 or clusPhiSize_ x clusEtaSize_)
-      sort(seeds.begin(), seeds.end(), ecalRecHitGreater);
+      std::sort(seeds.begin(), seeds.end(), ecalRecHitGreater);
 
       for (std::vector<EcalRecHit>::iterator itseed = seeds.begin(); itseed != seeds.end(); itseed++) {
         EBDetId seed_id = itseed->id();
@@ -424,8 +703,8 @@ void DQMSourcePi0::analyze(const Event &iEvent, const EventSetup &iSetup) {
         // Reject the seed if not able to build the cluster around it correctly
         // if(clus_v.size() < clusEtaSize_*clusPhiSize_){cout<<" Not enough
         // RecHits "<<endl; continue;}
-        vector<EcalRecHit> RecHitsInWindow;
-        vector<EcalRecHit> RecHitsInWindow5x5;
+        std::vector<EcalRecHit> RecHitsInWindow;
+        std::vector<EcalRecHit> RecHitsInWindow5x5;
 
         double simple_energy = 0;
 
@@ -522,7 +801,7 @@ void DQMSourcePi0::analyze(const Event &iEvent, const EventSetup &iSetup) {
         if (e3x3 <= 0)
           continue;
 
-        float s4s9_max = *max_element(s4s9_tmp, s4s9_tmp + 4) / e3x3;
+        float s4s9_max = *std::max_element(s4s9_tmp, s4s9_tmp + 4) / e3x3;
 
         /// calculate e5x5
         std::vector<DetId> clus_v5x5 = topology_p->getWindow(seed_id, 5, 5);
@@ -596,7 +875,7 @@ void DQMSourcePi0::analyze(const Event &iEvent, const EventSetup &iSetup) {
                                (p0y + p1y) * (p0y + p1y) - (p0z + p1z) * (p0z + p1z));
             if ((m_inv < seleMinvMaxPi0_) && (m_inv > seleMinvMinPi0_)) {
               // New Loop on cluster to measure isolation:
-              vector<int> IsoClus;
+              std::vector<int> IsoClus;
               IsoClus.clear();
               float Iso = 0;
               TVector3 pairVect = TVector3((p0x + p1x), (p0y + p1y), (p0z + p1z));
@@ -696,22 +975,22 @@ void DQMSourcePi0::analyze(const Event &iEvent, const EventSetup &iSetup) {
       // cout<< " Eta seeds: "<<seeds.size()<<endl;
 
       int nClus;
-      vector<float> eClus;
-      vector<float> etClus;
-      vector<float> etaClus;
-      vector<float> thetaClus;
-      vector<float> phiClus;
-      vector<EBDetId> max_hit;
+      std::vector<float> eClus;
+      std::vector<float> etClus;
+      std::vector<float> etaClus;
+      std::vector<float> thetaClus;
+      std::vector<float> phiClus;
+      std::vector<EBDetId> max_hit;
 
-      vector<vector<EcalRecHit>> RecHitsCluster;
-      vector<vector<EcalRecHit>> RecHitsCluster5x5;
-      vector<float> s4s9Clus;
-      vector<float> s9s25Clus;
+      std::vector<std::vector<EcalRecHit>> RecHitsCluster;
+      std::vector<std::vector<EcalRecHit>> RecHitsCluster5x5;
+      std::vector<float> s4s9Clus;
+      std::vector<float> s9s25Clus;
 
       nClus = 0;
 
       // Make own simple clusters (3x3, 5x5 or clusPhiSize_ x clusEtaSize_)
-      sort(seeds.begin(), seeds.end(), ecalRecHitGreater);
+      std::sort(seeds.begin(), seeds.end(), ecalRecHitGreater);
 
       for (std::vector<EcalRecHit>::iterator itseed = seeds.begin(); itseed != seeds.end(); itseed++) {
         EBDetId seed_id = itseed->id();
@@ -733,8 +1012,8 @@ void DQMSourcePi0::analyze(const Event &iEvent, const EventSetup &iSetup) {
         // Reject the seed if not able to build the cluster around it correctly
         // if(clus_v.size() < clusEtaSize_*clusPhiSize_){cout<<" Not enough
         // RecHits "<<endl; continue;}
-        vector<EcalRecHit> RecHitsInWindow;
-        vector<EcalRecHit> RecHitsInWindow5x5;
+        std::vector<EcalRecHit> RecHitsInWindow;
+        std::vector<EcalRecHit> RecHitsInWindow5x5;
 
         double simple_energy = 0;
 
@@ -831,7 +1110,7 @@ void DQMSourcePi0::analyze(const Event &iEvent, const EventSetup &iSetup) {
         if (e3x3 <= 0)
           continue;
 
-        float s4s9_max = *max_element(s4s9_tmp, s4s9_tmp + 4) / e3x3;
+        float s4s9_max = *std::max_element(s4s9_tmp, s4s9_tmp + 4) / e3x3;
 
         /// calculate e5x5
         std::vector<DetId> clus_v5x5 = topology_p->getWindow(seed_id, 5, 5);
@@ -905,7 +1184,7 @@ void DQMSourcePi0::analyze(const Event &iEvent, const EventSetup &iSetup) {
                                (p0y + p1y) * (p0y + p1y) - (p0z + p1z) * (p0z + p1z));
             if ((m_inv < seleMinvMaxEta_) && (m_inv > seleMinvMinEta_)) {
               // New Loop on cluster to measure isolation:
-              vector<int> IsoClus;
+              std::vector<int> IsoClus;
               IsoClus.clear();
               float Iso = 0;
               TVector3 pairVect = TVector3((p0x + p1x), (p0y + p1y), (p0z + p1z));
@@ -982,7 +1261,7 @@ void DQMSourcePi0::analyze(const Event &iEvent, const EventSetup &iSetup) {
       std::vector<EcalRecHit> seedsEndCap;
       seedsEndCap.clear();
 
-      vector<EEDetId> usedXtalsEndCap;
+      std::vector<EEDetId> usedXtalsEndCap;
       usedXtalsEndCap.clear();
 
       ////make seeds.
@@ -1016,20 +1295,20 @@ void DQMSourcePi0::analyze(const Event &iEvent, const EventSetup &iSetup) {
       //      "<<rhEEpi0->size()<<" "<<etot/rhEEpi0->size()<<" "<<etot<<endl;
 
       int nClusEndCap;
-      vector<float> eClusEndCap;
-      vector<float> etClusEndCap;
-      vector<float> etaClusEndCap;
-      vector<float> thetaClusEndCap;
-      vector<float> phiClusEndCap;
-      vector<vector<EcalRecHit>> RecHitsClusterEndCap;
-      vector<vector<EcalRecHit>> RecHitsCluster5x5EndCap;
-      vector<float> s4s9ClusEndCap;
-      vector<float> s9s25ClusEndCap;
+      std::vector<float> eClusEndCap;
+      std::vector<float> etClusEndCap;
+      std::vector<float> etaClusEndCap;
+      std::vector<float> thetaClusEndCap;
+      std::vector<float> phiClusEndCap;
+      std::vector<std::vector<EcalRecHit>> RecHitsClusterEndCap;
+      std::vector<std::vector<EcalRecHit>> RecHitsCluster5x5EndCap;
+      std::vector<float> s4s9ClusEndCap;
+      std::vector<float> s9s25ClusEndCap;
 
       nClusEndCap = 0;
 
       // Make own simple clusters (3x3, 5x5 or clusPhiSize_ x clusEtaSize_)
-      sort(seedsEndCap.begin(), seedsEndCap.end(), ecalRecHitGreater);
+      std::sort(seedsEndCap.begin(), seedsEndCap.end(), ecalRecHitGreater);
 
       for (std::vector<EcalRecHit>::iterator itseed = seedsEndCap.begin(); itseed != seedsEndCap.end(); itseed++) {
         EEDetId seed_id = itseed->id();
@@ -1048,8 +1327,8 @@ void DQMSourcePi0::analyze(const Event &iEvent, const EventSetup &iSetup) {
         std::vector<DetId> clus_v = topology_ee->getWindow(seed_id, clusEtaSize_, clusPhiSize_);
         std::vector<std::pair<DetId, float>> clus_used;
 
-        vector<EcalRecHit> RecHitsInWindow;
-        vector<EcalRecHit> RecHitsInWindow5x5;
+        std::vector<EcalRecHit> RecHitsInWindow;
+        std::vector<EcalRecHit> RecHitsInWindow5x5;
 
         float simple_energy = 0;
 
@@ -1131,7 +1410,7 @@ void DQMSourcePi0::analyze(const Event &iEvent, const EventSetup &iSetup) {
         etaClusEndCap.push_back(clus_pos.eta());
         thetaClusEndCap.push_back(theta_s);
         phiClusEndCap.push_back(clus_pos.phi());
-        s4s9ClusEndCap.push_back(*max_element(s4s9_tmp, s4s9_tmp + 4) / e3x3);
+        s4s9ClusEndCap.push_back(*std::max_element(s4s9_tmp, s4s9_tmp + 4) / e3x3);
         s9s25ClusEndCap.push_back(e3x3 / e5x5);
         RecHitsClusterEndCap.push_back(RecHitsInWindow);
         RecHitsCluster5x5EndCap.push_back(RecHitsInWindow5x5);
@@ -1168,7 +1447,7 @@ void DQMSourcePi0::analyze(const Event &iEvent, const EventSetup &iSetup) {
 
             if ((m_inv < seleMinvMaxPi0EndCap_) && (m_inv > seleMinvMinPi0EndCap_)) {
               // New Loop on cluster to measure isolation:
-              vector<int> IsoClus;
+              std::vector<int> IsoClus;
               IsoClus.clear();
               float Iso = 0;
               TVector3 pairVect = TVector3((p0x + p1x), (p0y + p1y), (p0z + p1z));
@@ -1234,7 +1513,7 @@ void DQMSourcePi0::analyze(const Event &iEvent, const EventSetup &iSetup) {
       std::vector<EcalRecHit> seedsEndCap;
       seedsEndCap.clear();
 
-      vector<EEDetId> usedXtalsEndCap;
+      std::vector<EEDetId> usedXtalsEndCap;
       usedXtalsEndCap.clear();
 
       ////make seeds.
@@ -1268,20 +1547,20 @@ void DQMSourcePi0::analyze(const Event &iEvent, const EventSetup &iSetup) {
       //      "<<rhEEeta->size()<<" "<<etot/rhEEeta->size()<<" "<<etot<<endl;
 
       int nClusEndCap;
-      vector<float> eClusEndCap;
-      vector<float> etClusEndCap;
-      vector<float> etaClusEndCap;
-      vector<float> thetaClusEndCap;
-      vector<float> phiClusEndCap;
-      vector<vector<EcalRecHit>> RecHitsClusterEndCap;
-      vector<vector<EcalRecHit>> RecHitsCluster5x5EndCap;
-      vector<float> s4s9ClusEndCap;
-      vector<float> s9s25ClusEndCap;
+      std::vector<float> eClusEndCap;
+      std::vector<float> etClusEndCap;
+      std::vector<float> etaClusEndCap;
+      std::vector<float> thetaClusEndCap;
+      std::vector<float> phiClusEndCap;
+      std::vector<std::vector<EcalRecHit>> RecHitsClusterEndCap;
+      std::vector<std::vector<EcalRecHit>> RecHitsCluster5x5EndCap;
+      std::vector<float> s4s9ClusEndCap;
+      std::vector<float> s9s25ClusEndCap;
 
       nClusEndCap = 0;
 
       // Make own simple clusters (3x3, 5x5 or clusPhiSize_ x clusEtaSize_)
-      sort(seedsEndCap.begin(), seedsEndCap.end(), ecalRecHitGreater);
+      std::sort(seedsEndCap.begin(), seedsEndCap.end(), ecalRecHitGreater);
 
       for (std::vector<EcalRecHit>::iterator itseed = seedsEndCap.begin(); itseed != seedsEndCap.end(); itseed++) {
         EEDetId seed_id = itseed->id();
@@ -1300,8 +1579,8 @@ void DQMSourcePi0::analyze(const Event &iEvent, const EventSetup &iSetup) {
         std::vector<DetId> clus_v = topology_ee->getWindow(seed_id, clusEtaSize_, clusPhiSize_);
         std::vector<std::pair<DetId, float>> clus_used;
 
-        vector<EcalRecHit> RecHitsInWindow;
-        vector<EcalRecHit> RecHitsInWindow5x5;
+        std::vector<EcalRecHit> RecHitsInWindow;
+        std::vector<EcalRecHit> RecHitsInWindow5x5;
 
         float simple_energy = 0;
 
@@ -1383,7 +1662,7 @@ void DQMSourcePi0::analyze(const Event &iEvent, const EventSetup &iSetup) {
         etaClusEndCap.push_back(clus_pos.eta());
         thetaClusEndCap.push_back(theta_s);
         phiClusEndCap.push_back(clus_pos.phi());
-        s4s9ClusEndCap.push_back(*max_element(s4s9_tmp, s4s9_tmp + 4) / e3x3);
+        s4s9ClusEndCap.push_back(*std::max_element(s4s9_tmp, s4s9_tmp + 4) / e3x3);
         s9s25ClusEndCap.push_back(e3x3 / e5x5);
         RecHitsClusterEndCap.push_back(RecHitsInWindow);
         RecHitsCluster5x5EndCap.push_back(RecHitsInWindow5x5);
@@ -1420,7 +1699,7 @@ void DQMSourcePi0::analyze(const Event &iEvent, const EventSetup &iSetup) {
 
             if ((m_inv < seleMinvMaxEtaEndCap_) && (m_inv > seleMinvMinEtaEndCap_)) {
               // New Loop on cluster to measure isolation:
-              vector<int> IsoClus;
+              std::vector<int> IsoClus;
               IsoClus.clear();
               float Iso = 0;
               TVector3 pairVect = TVector3((p0x + p1x), (p0y + p1y), (p0z + p1z));
@@ -1506,3 +1785,5 @@ int DQMSourcePi0::diff_nphi_s(Int_t nphi1, Int_t nphi2) {
   }
   return mdiff;
 }
+
+DEFINE_FWK_MODULE(DQMSourcePi0);
